Add cloud sync to intro screen
