+++ lib/l10n/app_en.arb
@@ -780,8 +780,10 @@
     "downloadToAppSubtitle": "Attempt to download and view PDF directly in Wispar for better integration.",
     "loginToSyncDevices":"Login to sync across devices",
     "@loginToSyncDevices":{},
-    "syncDisclaimer": "Disclaimer: This feature is experimental and provided as-is, with no guarantees of availability or reliability.",
-    "@syncDisclaimer":{},
+    "syncLongDescription":"Wispar offers an optional sync service to keep your data consistent across devices. You can use Wispar Sync (see Privacy Policy) or self-host for full data control (see Documentation). You can also skip sync and continue.",
+    "@syncLongDescription":{},
+    "documentation":"Documentation",
+    "@documentation":{},
     "cloudSync": "Cloud sync",
     "@cloudSync":{},
     "selfHosted":"Self-Hosted",
@@ -826,6 +828,8 @@
     "@syncSuccess":{},
     "login":"Login",
     "@login":{},
+    "userIsLoggedIn":"You are now logged in! Your data will stay in sync.",
+    "@userIsLoggedIn":{},
     "logout":"Logout",
     "@logout":{},
     "user":"User: {name}",
